+++ libs/components/src/lib/directives/hint/hint.directive.ts
@@ -131,7 +131,10 @@ export class PrizmHintDirective<
 
   public ngOnInit(): void {
     this.initVisibleController();
+    this.initShowedChangeListener();
+  }
 
+  protected initShowedChangeListener() {
     this.prizmHoveredChange$$
       .pipe(
         distinctUntilChanged(),