+++ libs/components/src/lib/directives/tooltip/tooltip.directive.ts
@@ -73,8 +73,11 @@ export class PrizmTooltipDirective extends PrizmHintDirective {
 
   protected override readonly containerComponent = PrizmTooltipContainerComponent;
   protected override readonly onHoverActive = false;
-
+  protected clickedInside = false;
   @HostListener('document:click', ['$event.target']) public onClick(target: HTMLElement): void {
-    this.show$.next(this.elementRef.nativeElement.contains(target));
+    const clickOnTooltip = this.elementRef.nativeElement.contains(target);
+    if (clickOnTooltip && !this.clickedInside) this.clickedInside = true;
+    if (!this.clickedInside) return;
+    this.show$.next(clickOnTooltip);
   }
 }